first

+++ database/migrations/2020_05_19_194259_create_users_roles_table.php
@@ -14,7 +14,7 @@ class CreateUsersRolesTable extends Migration
     public function up()
     {
         Schema::create('users_roles', function (Blueprint $table) {
-            $table->unsignedSmallInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('role_id');
 
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
