+++ database/migrations/2020_05_19_224316_create_users_permissions_table.php
@@ -14,7 +14,7 @@ class CreateUsersPermissionsTable extends Migration
     public function up()
     {
         Schema::create('users_permissions', function (Blueprint $table) {
-            $table->unsignedSmallInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('permission_id');
 
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
